+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 1. New highlighter color for: light, dark, ocean, dark-blue themes
 2. Fixed autocompletion templates in editor
 3. Fixed documentation bug
+4. Github API connection errors
 
 # Version 2.2