+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 2.6.0
+
+### Fixed/Changes:
+
+1. Improved github repo structure
+2. Fixed autocompleter insertion
+
 # Version 2.5.0
 
 ### Major updates